[resultdb] Batch TestResult writes in DeriveInvocation.
Spanner limits the number of mutations per transaction, and
a Swarming task may contain more test results than allowed.

Batch the TestResults and associate them with an Invocation
that behaves as container. After writing all the batches we
can include their container Invocations inside the original
Invocation associated with the Swarming task to be derived.

BUG=1027194

Change-Id: I0167f871f23345ec109bb3abfb1436ded3430c7d