[resultdb] Remove invocations by tag
Query RPCs currently can retrieve results from multiple invocations and
merge their results. This semantics is equivalent to having querying
one invocation that includes all matches invocations.

This semantics does not make sense for tagged invocations. Tagging
is useful to find invocations having some common characteristic, but
it does not make sense to merge their results because they aren't
a part of one computation.

Another problem is that BigQuery export messages currently happen to
accept an InvocationPredicate which does not make sense.

Indexing of invocations by tags was designed when we were not sure
if CQ will implement their own storage/API for CQ attempts. As of now,
the decision is that CQ will provide APIs, so we no longer have a use
case for indexing by tags.

This CL is a relatively large refactoring that solve the above problems:
1) Removes indexing of invocations by tags entirely. We can always
   restore this code later
2) Replace InvocationPredicate with just a list of invocation names,
   and moves it from XXXPredicate to QueryXXXRequest.
3) Removes the ability to ignore inclusions. No use case. We can add
   it later.
4) Introduces a helper type InvocationIDSet and uses it through the
   code base.

Change-Id: I0c671dcaac02a3139f4b5e14b021c769af22150a