[server] Add minimal tracing framework.
LUCI libraries can now use trace.BeginSpan(...) to report trace spans.

When tracing is not enabled (e.g. on GAE or in client-side binaries), it has
almost zero overhead and brings zero dependencies.

When executing in a server.Server environment, traces are collected and uploaded
to StackDriver via OpenCensus.io library (which is already a dependency of
server.Server).

Annotate Authenticate(...) as an example and a test.

R=tandrii@chromium.org, nodir@chromium.org

Change-Id: I3d623044c74e5aa0acea1e660bc596b0548c5434