[resultdb] Except base values in swarming task def
Currently DeriveInvocation accepts base test variant and test path
prefix in the request. This precludes prewarming of cache, and frankly
makes it harder for recipes to use this API.

Instead, (re)use magical tags in a swarming task, e.g. expect builder name
to be "buildername" tag, which is already there.
This means taskwatcher will prewarm the cache and DeriveInvocation will
finish from the perspective of the recipe.

Change-Id: I62273721303f3767f71e03fb73d7b4c2c6f6a9a4