[resultdb] Remove invocations by tag
Query RPCs currently can retrieve results from multiple invocations and
merge their results. This semantics is equivalent to querying one
invocation that includes all matched invocations.

This semantics does not make sense for tag search. Tagging
is useful to find invocations having some common characteristic, but
it does not make sense to merge their results because they aren't
a part of one computation.

Another problem is that BigQuery export messages currently happen to
accept an InvocationPredicate which does not make sense.
BigQuery export must be limited to the invocation we are exporting.

Indexing of invocations by tags was designed when we were not sure
if CQ will provide an API to fetch CQ attempts. As of now,
the decision is that CQ will provide such API, so we no longer have
a use for indexing by tags.

This CL is a relatively large refactoring that solves the above problems:
1) Removes indexing of invocations by tags entirely. We can always
   restore this code later
2) Replace InvocationPredicate with just a list of invocation names,
   and moves it from XXXPredicate to QueryXXXRequest.
3) Removes the ability to ignore inclusions. No use case. We can add
   it later.
4) Introduces a helper type InvocationIDSet and uses it through the
   code base.

R=chalni@chromium.org, jchinlee@chromium.org

Change-Id: I0c671dcaac02a3139f4b5e14b021c769af22150a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1944641
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>