[resultdb] Shorten swarming invocation ids
Invocation ids are duplicated in all test results, so it is benefitial
to make them shorter.
Also DerinveInvocationRequest no longer has input parameters, so
there is no need to compute hash of the request. It is incorrect
anyway because protobuf wire format is not deterministic.

Use simple "swarm:<hostname>:<runid>" format for invocation ids.

Bug: 1027922
Change-Id: Ib975abc355d8a68dc3e76d22d1d02cee3eaa58d3