[resultdb] Allow skipping processing some output files.
Some output files, such as profraw files, we know we may not process.
Do not log them as errors if they show up as unprocessed outputs.

In any case, note file names and file prefixes we know may not be
processed and don't need to be in the first iterations of resultdb,
even if we plan to process them later.

Change-Id: I75525d1fb9b118443b2afd48ce6efb1b0d0bf72c