cq: add Builder.includable_only option for `CQ-Include-Tryjobs` use only.
In the future, all builders which CQ is allowed to trigger within
the context of a project must be listed in the project's config.
Today, the only exception are builders freely added via
`Cq-Include-Trybots:` in CL description.

This CL adds a way to add such optional builders to a project config
s.t. CQ knows about them yet won't trigger by default.

R=vadimsh

Bug: 937080
Change-Id: I639423d210e68f090262f7bda8604fab950514b2