[resultdb] Replace GN label with GN target
GN label does not provide a guarantee that a given test sub-path is
unique within a GN label because a GN label can group multiple test
binaries together.

Also a GN label to include another label, which introduces ambiguity and
makes the test identity more volatile, a property that an identity
should avoid.

Also currently we accept "gn_target" and treat it as a label. There is
no such a thing as "GN target". There are only "GN label" and
"Ninja target".

Changes:
* Rename gn_target tag to ninja_target. Change the prefix from
  "gn:{gn_taret}" to "ninja/{target}"
* Update all docs and tests.
Change-Id: I309cc44ff94d67d6bf50054c70ebe2f98bc7bba1