[resultdb] Switch to Ninja
It turned out a GN label does not guarantee uniqueness of a test
sub-path because one GN label can include two Ninja targets.
The guarantee is provided only at the Ninja level.

Also the tag we parse is "gn_target", but "GN Target" is not a thing.
There are "GN labels" and "Ninja targets".

Replace "gn:" prefix with "ninja" and accept rename gn_target to
ninja_target.

Change-Id: Ic460e90a31fc289e0ab40e2c52496b761562a451