Update patch set 1
Patch Set 1: Commit-Queue+2

> Patch Set 1: Code-Review+1
> 
> chatted offline. What we are doing now is incorrect, but this CL improves the situation, so LGTM

Yes, this is definitely weird and inconsistent with the other places we distinguish the individual artifacts into runs, but making it consistent would require much more invasive adjustments unfortunately.

Fortunately this should happen rarely and when it does the links should probably even be the same (but the runs are retaining information about all the links so in the worst case the information is not lost).

Thank you!

Patch-set: 1
Reviewer: Gerrit User 1134874 <1134874@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+2