[resultdb] Drop Markdown
From buildbucket experience, Markdown proved to be a bad choice for
machine-generated rich text. Programs often need to deal with arbitrary
text that they include in their reports. This means this text must be
escaped. Escaping markdown caused quite a bit of problems in buildbucket
experience.

In practice, rich text is always converted to HTML, so just use HTML
instead.

Rename SummaryMarkdown and ExplanationMarkdown to just Summary and Explanation.
Not FooHTML, because then the proper full name would be FooUnsafeHTML, but that's
too long.

Bug: 1035777
Change-Id: Ib6ba23c32d15da4dcbdc54814161542449242ddd