[prpc] client: extract response body reading
prpc.Client.call is ~200 LOC, which is too much.
Extract 30 LOC of reading of the response body to a separate function.

Also fix a bug: the "probe" check was dead because it was reading from
overwritten, limited body variable which definitely would not return
any additional content.

Simplify logging: ditch fields.

Bug: 1036836
Change-Id: I1097001bebd64d57ce0c178855624b0be8f22696