[resultdb] Add appstatus package
grpcutil has a tag for gRPC code. It works fine in isolation. However,
with an addition of status details, it is not longer preferrable to
use isolated status code tag because code and details may be collected
from different errors.

Another slight problem with code tag is that that code that specifies
that code does not have full control over the error message, but rather
its suffix.

Another slight problem is application-specific code has to exclude
internal errors, e.g. statuses returned by Spanner client.

Reduce magic by adding a tag with an application-specific
*status.Status. Forbid overriding it in a chain of errors.

Clearly public and internal error messages: error messages are internal,
status messages are public. A decision to expose err.Error()
is explicit.

Add ShouldHaveAppStatus assertion to check an error for an
application-specific status.

Do not place appstatus package outside of resultdb until it proves
useful.

Convert all code. Usage of the new package make the code shorter.
This CL completely ditches the usage of grpcutil package.

Do not explicitly annotate errors with INTERNAL code. This is the default.

Use %s (not %q) for invocation name because it may not have spaces or
other weird characters.

Bug: 1035467
Change-Id: Ifcc3ef8814737bf0474dc7565c174fe673b9524b