[server] Add a simple limiter for number of concurrently handled RPCs.
The focus of this CL is to come up with a good API for the limiter (including
metrics!), so that we can teach it more tricks later without significantly
changing the callers.

Another caveat is that we install the limiter only on gRPC handlers for now
(since that's what we need). Installing the limiter on all HTTP routes on the
generic server.Server level is complicated because pRPC requests are also routed
through the default HTTP port and we want configs and metrics for pRPC requests
to be represented in terms of gRPC concepts, not in terms of HTTP requests.

R=tandrii@chromium.org, nodir@chromium.org
BUG=1033758

Change-Id: Ie014ed6d218d6483348d3c2f348925cd19dc19c5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1984803
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>