Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> It seems wrong to have both integers and strings. But if we remove integers, it will break some alerts (see https://cs.corp.google.com/search/?q=%22chrome/infra/grpc/server%22&sq=package:piper+file://depot/google3+-file:google3/experimental&type=cs). 
> 
> Scott, wdyt?

I'd love to remove the integer field, but we need to do it gradually, as we won't be able to update all the apps with the latest tsmon-go and monitoring dashboards in one-go. We need a migration plan like below.

1. Land this CL
2. Announce the change, and ask the owners of https://screenshot.googleplex.com/4bKHcxiS5r8 for updates.
3. Update all the monitoring alert and dashboard configs.
4. Land another CL to kill the integer field in tsmon-go.
5. Ask the owners again for another update.

Patch-set: 2
Reviewer: Gerrit User 1147264 <1147264@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>