Set kitchen's gitconfig template to use HTTP/1.1
The GoB team reports that they have seen errors when using
HTTP/2 to talk to googlesource through GFEs. We experience
these as hangs where the client remains connected to a
dead server process.

The version of libcurl.so found on GCE bots appears to be
the same as the version distributed to SWE gLinux desktops.
They have fixed the issue on google-managed machines by
enforcing the equivalent of
`git config --global http.version HTTP/1.1`
on workstations. We should do the same on our bots.

Bug: b/137034666
Change-Id: If468902e1e961ca7493be7acbfaf234e58dbf4fe