[auth] Cache results of /auth/api/v1/server/info in a two layer cache.
This is a prelude to a similar change for /auth/api/v1/server/certificates.

It is a resiliency measure: if a service goes down, we still want an ability
to verify signatures produced by it. Using only an in-process LRU cache means
we'll lose cached keys when the server restarts. A two-layered cache utilizes
an external cache that survives process restarts.

R=nodir@chromium.org

Change-Id: I669b9996a3b6aaf0100a0413e33ac1116a24a46d