[auth] Cache fetched service account public keys in a two layer cache.
It is a resiliency measure: if a service goes down, we still want an ability
to verify signatures produced by it. Using only an in-process LRU cache means
we'll lose cached keys when the server restarts. A two-layered cache utilizes
an external cache that survives process restarts.

Also bump the cache expiration time to 6 hours (it means we can potentially
survive more prolonged outages). Remove CertsCacheExpiration from the public
API, no one should depend on it.

R=nodir@chromium.org

Change-Id: I0805b33a48c93f646042e2990c5229a2b3906f59