[prpc] Handle HTTP 5xx statuses
If a client observes an HTTP 5xx without an explicit gRPC code, it is
possible that that request did not reach the pRPC server, e.g. it was
aborted by the hosting platform. It is still preferable to treat such
responses as INTERNAL/UNAVAILABLE errors in the pRPC client. Do that.

Change-Id: I05b239de19f75261bf62426e0409c41acc12557a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2001061
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>