[server] Add support for "virtual hosts".
Allows a single HTTP port to server different routers based on Host request
header. This is a feature which is not usually needed, so the API is optimized
for the case when this feature is NOT used.

In particular there are still server.Routes and server.PRPC which serve all
requests that do not match any registered virtual hosts. Usually there are no
virtual hosts, so they serve all requests, just like before this CL.

Note that if there's a need to expose pRPC on a "virtual host", it has to be
done explicitly via srv.PRPC.InstallHandlers(srv.VirtualHost("..."), ...).

R=nodir@chromium.org

Change-Id: Ida6737661fc21e2e891020bf802573fb6a3025db