[resultdb] simplify the argument list of testutil.InsertInvocation
There are some columns (for example Realm or Interrupted) in Invocation that most calls to InsertInvocation
only set them to default values. So it doesn't really make sense to grow the argument list just to accommodate
one or two special calls. In this change, move those arguments into a map.

Change-Id: Ied2289d91f9e9e199b30c9e3c1edadcd05df36ee