[resultdb] Simplify invocation task PK
Finalizing task needs to insert a task for multiple invocations.
It is inefficient to do so if InvocationTasks PK starts with invocation
id because tasks would be inserted into different parts of the keyspace.

InvocationTasks PK starts with InvocationID because initially it was
designed to be interleaved, but this is no longer the case.

Simplify the PK to be only the task id. Make InvocationID a regular
column. This simplifies the code too.

Bug: 1032434
Change-Id: I63666651ebac7d64e512f78fadbe2a30ed97628a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2009020
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>