[resultdb] Use shorter hash
Invocation id is copied in each test result row. It is currently
prefixed with a first 8 bytes of sha256 hash. The single purpose of the
prefix is even distribution across the key space. 4 bytes is more than
enough for this specific purpose, assuming sha256 is a good hash func.
Reduce the prefix length from 8 to 4 bytes.

R=vadimsh@chromium.org

Change-Id: If15e1a6066b216a844ee40eca061d2355eea7831