[server] Make authentication work with GAE v2 metadata server.
It returns cached tokens even if they are very close to expiration
(the limit is not documented, but appears to be ~1 min). This trips
auth.Authenticator which assumes that the token refresh procedure
(invoked when the token has about 2 min to live) actually refreshes
the token.

Make '2 min' constant configurable, change it to 20 sec on GAE v2.
Due to the refresh randomization feature of auth.Authenticator, it
implies that we may start attempting to refresh the token when it
has about 50 sec to live.

R=iannucci@chromium.org, nodir@chromium.org

Change-Id: I5a1f86d879734ac2c00cb11cfbde51a03a363e72