[resultdb] Add task type
We need to define different thresholds for different types of tasks,
so we need "type of task" as a first class citizen in the generic
task-related code.
Add TaskType as the first column in the InvocationTasks PK.
This is consistent with the future direction (Spanner queues) where
we'd have separate queues for different types of tasks.

Change-Id: I2036ac6888ad0d25f348e5d8e4422d7d1e28c470