[server] Make Cloud Tracing work on GAE.
This is a prerequisite for making logging work, since
Cloud Logging uses trace IDs to group logs together.

On GKE we used to always open a new root trace and link it
to the parent via AddLink (it's unclear whether it ever worked).

But GAE wants us to inherit the trace it opens, but adding
spans to it as children. This also works on GKE, so we do
it everywhere now.

On GAE we also let the appengine make sampling decisions
for us, so traces we collect are always full (include spans
produced by GAE itself), and sampling strategy matches whatever
it is on GAE v1.

R=iannucci@chromium.org, nodir@chromium.org

Change-Id: I90b9be01a24720ea992db2e99cbd51775429cd30