[resultdb] Add task type
We need to define different thresholds for different types of tasks,
so we need "type of task" as a first class citizen in the generic
task-related code.
Add TaskType as the first column in the InvocationTasks PK.
This is consistent with the future direction (Spanner queues) where
we'd have separate queues for different types of tasks.

R=chanli@chromium.org

Change-Id: I2036ac6888ad0d25f348e5d8e4422d7d1e28c470
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2015200
Commit-Queue: Chan Li <chanli@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>
Auto-Submit: Nodir Turakulov <nodir@chromium.org>