[server] Make logging work on GAE v2.
Also rename gkelogger module to sdlogger module ("StackDriver logging"), since
it is now used not only on GKE.

GAE's implementation of JSON log parsing is mostly, but not quite, similar to
google-fluentd. It appears to support a subset of fields/formats understood by
the fluentd plugin. In particular, it is picky about allowed severity levels
and likes timestamps to be represented by structs.

We switch to this subset, so JSON formatting works on both GAE and GKE as is.

Another difference is that GAE emits the root log request record (the one we
group other log entries against) itself, so we don't need to do it from our
code.

logging.SetField doesn't really work yet (fields end up being invisible in the
log viewer when looking at the grouped logs). I suspect this is also the case
for GKE. This will be fixed separately.

R=iannucci@chromium.org, nodir@chromium.org

Change-Id: I70c0e2e8feb3e58e325563684332ad83c018d6b4