[config] Mark packages we want to delete as deprecated, outline the plan.
The existing LUCI Config library is vast and poorly understood. Migrating it
to work on GKE will require substantial effort. Instead we'll try to replace
it with something simpler which covers 90% of users needs and promotes health
practices (like not to depend on the availability of the control plane from
the data plane).

R=tandrii@chromium.org, nodir@chromium.org, iannucci@chromium.org

Change-Id: Ie307f1dbf42aa999ca9ae2890a4043bb88a11227