Update patch set 2
Patch Set 2:

PTAL.

I propose to keep (but maybe move elsewhere):
 * Most of *.go directly under https://chromium.googlesource.com/infra/luci/luci-go/+/refs/heads/master/config/
 * All of https://chromium.googlesource.com/infra/luci/luci-go/+/refs/heads/master/config/impl/
 * All of https://chromium.googlesource.com/infra/luci/luci-go/+/refs/heads/master/config/validation/

Everything else will be replaced with an interface semantically similar to 

  config.Subscribe("<config set patterns>", "<files patterns>", callback)

And the framework will take care of calling `callback` once in a while. It will be users job to persist configs somewhere (like most of them already do in various cron jobs). 

(Accidentally, we need such patterns anyhow to setup config validation).

Then on GAEv2 this can be implemented via a simple cron job, and on GKE it will be a simple 1 replica pod with a loop (or something like that).

The main problems is regutting Logdog and Milo to not depend on "advanced" features cfgclient (like ACLs, wat).

Patch-set: 2
Work-in-progress: false