Update patch set 2
Patch Set 2:

PTAL and tell me if you hate it.

I resisted this idea for quite a while, but server.go is getting ridiculous...

I started thinking how to add integration with LUCI Config validation framework (it is always turned on on GAE), and the least resistance path is to just add it to server.go (so it is also always "turned on"). But it makes server.go the ultimate dumpster of everything.

So we need to keep the integration with validation framework elsewhere. This leads to a question how to hook it up to the server. I decided we need a generic way of "hooking up" optional things. And this CL proposed (part) of the API for this.

Tomorrow I'll try to split out some existing chunks of server.go into modules to get a feel for it.

Patch-set: 2
Work-in-progress: false