[cipd] Handle ScanFileSystem(<symlink>, <same symlink>) gracefully.
It happens when a symlink is passed to e.g. "cipd create -in ...".
It shouldn't matter that it is a symlink (the root of the package is
not actually a part of the package, only files under it are).

Before this CL an attempt to do so resulted in "Failed to build the
package: Invalid symlink .: a relative symlink pointing to a file
outside of the package root". Which is technically correct, but not
what one would expect.

Now we handle this better by traversing the symlinks before doing
filepath.Walk. This necessitates switching ScanFilter predicate to
use relative paths (since absolute paths now may include resolved
symlinks and be surprising to ScanFilter implementers). But the
only implementation of ScanFilter did the conversion to relative
paths anyway, so this change is easy.

Finally, cleanup some comments and fix minor nits in tests.

R=iannucci@chromium.org

Change-Id: I86b234946fce8df5e0af8d422312d7dd7bfe21f4