[config] Delay validation of rules in a RuleSet.
We relied on an import order of go packages to do rules.RegisterVar
and rules.Add in a correct sequence. This becomes complicated when
the config validation is moved into a separate module.Module which
is instantiated only when needed and generally *not* during init()
time.

This CL removes the order requirement. Vars and rules can be
registered in any order as long as everything is consistent when
rules.Freeze() is called, which will be called from module.Initialize.

R=tandrii@chromium.org, nodir@chromium.org

Change-Id: Ie64c33a771afafce01ed5a20e24b811e580879e0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2026605
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>