Update patch set 2
Patch Set 2:

Note that this changes tsmon target schema (that's the reason we no longer need the cron). It means some of our Viceroy consoles will have to be adjusted once we start using it for other apps.

The good news is that once adjusted, such consoles should also work for GKE-based services.

Patch-set: 2