[bb] Support raw BuildPredicate and arbitrary FieldMask for ls cmd
This change will allow bb user to pass in multiple buildPredicate proto
messages in JSON string so that we can support all the existing and future
predicates defined in the buildPredicate. Before only builder, tags,
gerritChanges, includeExperimental and status are supported.

This change will also allow bb user to declare any fields they need for each
build in the response.

R=nodir, tandrii

Bug: 1039823
Change-Id: Id8c42cce435ec20294817fb9e26c4d41de361cff