isolate: implement check
python's isolate.py generates isolated file in addition to check file
existence, but that will be done in archive command when we switched to
go client.
So does not generate isolated file in `check` subcommand in this CL.
`mb.py isolate` itself does not require isolate to generate isolated
file.
https://cs.chromium.org/chromium/src/tools/mb/mb.py?l=219&rcl=4b921433e0cd6da517a16d6127b5b78b77252fee

Bug: 1015837
Fixes: 1015837
Change-Id: Id6e9734f691424af020f6ee8ac562f66f3088689