[bb] Print build correctly when field mask is applied to build
The print build function will error out if the `builderId` field is
missing in build. As new `field` flag is introduced into `bb ls` command,
the return build may not contain such field. This change is to improve
the print build function to handle the scenario where certain fields are
missing.

R=nodir

Bug: 1039823
Change-Id: I0aee9850a2364f779316283ac8f0e80ad6034b0e