[bb] Print build correctly when field mask is applied to build
In the existing implementation, the print build function will error out
if the `builderId` field is missing from the given build. As new `fields`
flag is introduced to `bb ls` command, the returned builds may not have
such field. This change is to improve the print build function to handle
the scenario where certain fields are missing from the given build.

R=nodir

Bug: 1039823
Change-Id: I0aee9850a2364f779316283ac8f0e80ad6034b0e