[resultdb] Refactor bqexport
Reduce number of queryTestResultsStreaming parameters by making it
a method of bqExport.

Do not copy entire bqExport each time a call its method by
using a pointer receiver. This is a common practice.

Inline constants that are used in one place. Local variables in
queryTestResultsStreaming were shadowing the constants, which makes
the code harder to reason about. This consolidates backend tuning
in one place (backend.InitServer).

Change-Id: Ia6edbdddb0e0535a411f4e6cd32bf9076b38ccc3