[resultdb] Remove cron iteration timeout
It is challenging to come up with a good default for cron iteration
timeout. We lease 100 tasks and each has up to 10m. Which means a whole
iteration can take up to 1000 minutes which is more than an hour.

Remove cron iteartion timeout. If a particular cron job really needs it,
it can limits its own context itself.

R=chanli@chromium.org

Change-Id: If1ceef96eb42b0e25d12d77a135d237f64cf7798