[resultdb] Simplify cron function.
We started cron jobs in resultdb with short cron jobs, like on GAE, but
then switched to long jobs, where each job allocates ~100 work items and
processes it sequentially. With such jobs, minInterval is minutes, not
seconds, and there is no need in linear/exponential backoff.
Remove this and minSleep logic. It is no longer needed.

R=chanli@chromium.org, ddoman@chromium.org

Change-Id: Idb295257c6a6937d9f6be171f2688e342bb99a68