[resultdb] Extract purger
Test result purging and task processing are in the same deployment,
sharing the same number of replicas. However, they are very different:
purger spends very little CPU, most of the work is done by Spanner.
Task processing does much more CPU intensive work. If the number of
tasks to process grows, we have more replicas of purger unnecesarily.
If we really have a lot of test results to purge, we need to increase
the number of spanner replicas, not purger.

Split purger to a separate deployment, with replicas=1.

Do not purging code from backend yet, so that we don't have purgingdowntime

Change-Id: I9204f28104177778bafc724a7fd78252350a9580