cq: Add FailureType to Attempt proto
Noted by martiniss, in some current usage of cq_attempts,
rows are filtered on fail_type. In order to get an equivalent
result from the new table, some more details about failure
reason need to be surfaced.

This proposal is to just add a failure type field to Attempt,
which would be directly mappable from the old table.

Bug: 804289
Change-Id: I7cbb4a72e65d0bdc686e2131a91a0c83391ccb8a