[resultdb] Add to-purge invocation count metric
Every time purger is behind, it is difficult to see how much is left
to do from "oldest expired result" metric because its changes are not
smooth. So I tend to manually run a query to count number of pending
invocations. Automate this by adding a metric and computing it in
globalmetrics service.

It does make the query more expensive, but we run it only once in 5m,
and

Change-Id: If289f35b54ac299ffc4958472ef968ce073e674a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2067178
Reviewed-by: Roberto Carrillo <robertocn@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>