[resultdb] update lucictx resultdb with the latest schema
As described in "environment" section of http://go/result-sink

This CL renames the existing fields in a backward-incompatible way, but
the message struct of ResultDB has been publushed in the following document.
https://github.com/luci/luci-py/blob/master/client/LUCI_CONTEXT.md

Please let me know if this should keep the message struct backward-compatible.
Also, it renames the top-level key from "resultDB" to "resultdb". Looking at
other luci context go modules, such as local_auth, all the keys are made with
lowercase letters and a underscore, optionally.

R=nodir@chromium.org
Bug:1017288
Change-Id: I86cc611778bccb15e8b165c964401dc321c1ad83