Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1
> 
> lgmt, I guess, since I doubt we would ever want to store 1970 date, but the similar code in protobuf package handles it differently: https://github.com/golang/protobuf/blob/master/ptypes/timestamp.go#L90 (IIUC, zero Timestamp would result in 1970-01-01 00:00:00 +0000 without any errors). Would it be hard to indicate absence of the timestamp by the absence of the field (so 'nil' code path is hit)?

Ack.
I did fix the python round trip in our recipes which was replacing the nil with the zero value, but as the unittest here states, it is very easy to do the wrong thing when using python bindings. Even reading the nil value converts it to zero. You then write the proto back ....

Patch-set: 2