[resultdb] SinkServer is an HTTP server
(1) Server serves HTTP requests
As the decision was made in go/result-sink, Server is going to serve pRPC
requests, and this CL reimplements Server as an HTTP server.

(2) Run and Start no longer takes a context as a param
Let's say that a server was created by NewServer with context-Foo, but started
with context-Bar. It's confusing that the context used to create the server
context has no impact on the request serving context.

In this CL, the context passed onto NewServer is used as the serving context.

Change-Id: Ic4a19a82211203eb6454219c5f546900e6730be6
R=nodir@chromium.org
Bug:1017288