[auth] Workarounds for slow gke-metadata-server.
When there are many pods on a single GKE node, or they are making
many concurrent requests, gke-metadata-server pod becomes overloaded.
It results in metadata.OnGCE mistakenly reporting false and
ComputeTokenSource failing to produce a token.

This CL avoids metadata.OnGCE when it's not clear whether it would
work or not, replaces ComputeTokenSource with direct usage of
a metadata.Client with higher timeouts, adds mutex around it, and
tweaks various timeouts.

R=iannucci@chromium.org

Change-Id: I14f714aea45a7783a7ecb91b5c4ce2647e795de9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2084078
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>