Luci-notify should not return 500 when logdog query fails
We are having a high error rate of 500 error of POST request for luci-notify from pubsub.

When luci-notify gets a pubsub notification, one of the step is that it will query Logdog for the source manifest of the build. From the source manifest, it will get the revision for the build, and based on this revision, and the old revision it stores in datastore, it will attempt to calculate the blamelist.

Basically, the sole reason for querying logdog is to calculate the blamelist, in case notify_blamelist was turned on.

There are (many) cases where logdog cannot provide source manifest, because there is no source manifest for the build.

As blamelist is not a critical part of luci-notify, I don't think it should throw 500. Instead, it should continue to notify its recipient list. We can log the error somewhere and inform the build owner later (this is not handled in this CL, I created a bug for it, in case we still want to query from Logdog).

On a side note, I don't know if querying LogDog is the right way to calculate blamelist (e.g. Milo calculates its own blame list without Logdog), also I don't know if there exists something that we can query for blamelist instead of calculating it. However, we can decide later. This CL is just to get rid of the 500's relating to logdog, so troopers won't be paged.

Bug:1056842
Change-Id: I5143fdb663a79e45ed908f443d462bb982d0edfc