Correct timeout for logdog
This CL is a follow-up CL of https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2086412

The error rate in luci-notify did not lower with the deployment of the previous CL.

I think the reason is that in main.go, we set the context timeout of pubsub POST handler to be 50s. Logdog "inherits" this timeout and throws an error after 50s.

Although we ignores this error in the previous CL, and proceeds to other steps, but the context already times out, so if one of the subsequent steps check for this context's deadline, it will see that context has timeouted and bailed out, returning errors.

Let's fix it by having an overall timeout of pubsub POST request, and another (smaller) timeout for logdog.

Bug:1056842
Change-Id: I995808e1ac2c06f6fc4c3107948db20be6c214b7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2089185
Commit-Queue: Quang Minh Tuan Nguyen <nqmtuan@google.com>
Reviewed-by: Owen Rodley <orodley@chromium.org>