[led] Add pure luci-go main executable for `led`.
This version doesn't have kitchen support, but that's OK; Once we drop
kitchen, this version will be the one that lives on.

infra.git will have a different entrypoint which supplies kitchen
functionality (since it needs to link against 'infra/*' libraries to
parse the kitchen cook flags). We'll ship the infra.git version as the
first replacement during the transition.

R=smut@google.com, tandrii@chromium.org

Change-Id: I7bd89fb24ad859337dc5ed616b664ae64411f497
Bug: 1017818