[logdog] Restrict and simplify Query RPC implementation.
This restricts the logdog Query API to REQUIRE a prefix (not a wildcard),
and reimplements the Query filtering logic as in-memory over LogStream
entities returned from a basic query. This covers all the known uses
of the Query API.

For now this leaves the indices and field generation in LogStreams
intact so we can roll this out to LogDog instances as a canary without
worrying about data corruption if we need to roll back.

R=tandrii@chromium.org, vadimsh@chromium.org

Bug: 1061187
Change-Id: Iad010883606ab742f48788405b32842a47a3dfc3