Update patch set 3
Patch Set 3:

> Patch Set 3: Commit-Queue+1
> 
> (44 comments)
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > (22 comments)
> > > 
> > > is there a way to split this CL? It seems it tries to combine a lot of functionality: test result loading, hierarchy building. I didn't finish reviewing.
> > 
> > +1. I recommend having a read of go/small-cls if you haven't seen it before, it'll make the review process much faster and easier for everyone to manage.
> > 
> > Here's one way you could split this up:
> > CL #1: TestNode + unit tests
> > CL #2: TestLoader + unit tests
> > CL #3: UI components
> > CL #4: Adding the UI components to the invocation page
> 
> I see. I will break the next CL to smaller pieces.

I still recommend breaking up this one. It'll not only make the process go faster, but will allow us to do more thorough reviews and result in better code quality.

It's hard to do a thorough code review for a CL that adds almost 1000 new lines.

It should be reasonably easy. I'd do it by:
Leaving this CL as-is for now
Copying test_node.ts and test_node_test.ts into a new branch and mailing it to us as a new CL
Afterwards, you can either wait for that to be submitted, or you can create a new branch on top of it using:

$ git new-branch --upstream <parent-branch-name> <child-branch-name>

Then copy test_loader.ts and test_loader_test.ts into the new branch, upload it, and so on.

Once the CLs for test_node, test_loader and test_nav_tree are submitted, just rebase this one and only the invocation_page integration stuff should be left.

Patch-set: 3
Reviewer: Gerrit User 1272456 <1272456@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>