[gerrit]Modify gerrit cli output error info
we want to be able to detect when we should retry a gerrit operation. That's currently impossible because the CLI doesn't produce structured output when an API call fails. Output should include an error message field and boolean "transient" field. In the case of the roll-commenter, we can use these to determine what action to take in case a comment fails
Bug: fuchsia:47289