[gerrit]Modify gerrit cli output error info
we want to be able to detect when we should retry a gerrit operation. That's currently impossible because the CLI doesn't produce structured output when an API call fails. Output should include an error message field and boolean "transient" field. In the case of the roll-commenter, we can use these to determine what action to take in case a comment fails
Bug: fuchsia:47289
Change-Id: I16a1704a8fa9025c023e7d1bcafb6941cdbe9f6f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2113291
Commit-Queue: Mahesh Saripalli <maheshsr@google.com>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>