Move filtering code from api/config into notify/
Since it's no longer in the same package, we also have to change these
from methods on the proto types to just free functions taking them as
arguments.

Rationale:
This is partially because we want to re-use the step filtering code for
tree closing, and it's awkward when it's in this package. And partially
just because the existing structure feels strange. This code is
conceptually related to notifying, so it feels odd to define it as
methods on the proto type, within the proto package.

It makes more sense to me to keep api/config as purely proto
definitions, and have this code in notify/ along with the rest of the
related code.

Bug: 1051371
Change-Id: I303954162addb2339316a0b5292dd63ecdb03690
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2119471
Reviewed-by: Joey Scarr <jsca@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Auto-Submit: Owen Rodley <orodley@chromium.org>
Commit-Queue: Owen Rodley <orodley@chromium.org>