[logdog] Refactor archivist to form a pipeline.
Currently the archivists consume a batch (~500 tasks) from all
available tasks. The VAST majority of tasks are very tiny logs. This
means that archivits which consume a batch of 499 tiny logs and 1 big
log will end up doing almost nothing while the 1 big log completes.

This CL refactors the archivist into a pipeline consisting of three phases:
  * A leaser goroutine
  * An ACKer goroutine
  * A workpool to transform leases into ACKs (i.e. do the work)

The leaser feeds leased tasks into a channel for the workpool to consume;
if the workpool has workers ready, the leaser will fetch another batch,
even if some workers are still processing a log or two.

The ACKer goroutine will buffer ACKs until they're 30s from expiration,
or it accumulates 500 ACKs, whichever comes first. The ACKing process
doesn't block anything (ACKs are really an optimization in the grand
scheme of things; an ACKable task is already recorded in the datastore
as being archived and the coordinator will not re-issue it).

The work pool is a simple concurrent pool with a configurable size;
Unlike previously, the concurrency setting for the archivers will now
actually be relevant ~100% of the time the archiver is working, because
the workers will never starve.

R=tandrii@chromium.org, vadimsh@chromium.org

Bug: 1061187
Change-Id: I07b25a566f5479945f50db55a796bfc20c82b538