Update patch set 13
Patch Set 13:

> Icon issue: Added cursor: pointer. Not sure about changing the icon to '>', it's already used in expanding entries. What about https://material.io/resources/icons/?icon=filter_list&style=baseline

Ack, let's leave it as-is for now and we can figure out the best affordance later.

> > Can we also make it clearer 'why' tests won't be loaded.
> Not sure what you meant here. Could you clarify?

I find the "load more" with no context confusing and anxiety-inducing (as a user), because it doesn't explain to me what's missing or how much data is missing.

Why are tests not loaded? Is it because of pagination, or because we're only showing unexpected results? How many tests are not loaded? The UI explains none of these things.

Could we say something like "Showing X out of Y test results. LOAD MORE"?

Patch-set: 13
Reviewer: Gerrit User 1272456 <1272456@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>